Time: 1 ms (99.85%) | Memory: 41.9 MB (99.38%) - LeetSync

+++ 4168-mirror-distance-of-an-integer/mirror-distance-of-an-integer.java
@@ -8,6 +8,6 @@ class Solution {
             temp /= 10;
         }
 
-        return (int)(Math.abs(n - rev));
+        return Math.abs(n - rev);
     }
 }
